requirements.txt
aniso8601==4.1.0
astroid==1.6.5
awscli==1.16.64
backports.functools-lru-cache==1.5
botocore==1.12.54
certifi==2018.11.29
Click==7.0
colorama==0.3.9
configparser==3.5.0
docutils==0.14
enum34==1.1.6
Flask==1.0.2
Flask-RESTful==0.3.7
futures==3.2.0
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
pyasn1==0.4.4
pylint==1.9.3
python-dateutil==2.7.5
python-dotenv==0.10.1
pytz==2018.9
PyYAML==3.13
rsa==3.4.2
s3transfer==0.1.13
singledispatch==3.4.0.3
six==1.11.0
squareconnect==2.20181212.0
urllib3==1.24.1
virtualenv==16.0.0
Werkzeug==0.14.1
wrapt==1.10.11
